makefile: Change from bash script to Makefile system
This is much more flexible system, with the following (main) tasks:

  - build (mac64, default)
  - clean
  - install
  - uninstall
  - all (win32/64, linux32/64, mac64)

The version / `main.BUILD_TAG` will be composed like:

    PLATFORM-ARCH-YYYYMMDD-HHMMSS-SHA1
    windows-amd64-20150317-082841-da51745

This can be viewed in the header output of the command or (`-v`)

There are other specific or utility commands, the most
notable is `goinst`, which is invoked as the last step of `install`.

This allows us to test / use `go install` with a build tag without
updating dependancies. This is needed when building the project from
 a different branch because an update (`go get -u`) will switch back
 to the master branch.